Revert "Revert "RayTracing""
This reverts commit d3306aba5e8bae7d8ebf3002943d28f0324ad333.

+++ GameTest/src/edu/wkashti/game/Calculate.java
@@ -102,7 +102,7 @@ public class Calculate extends Main {
 		if (i == iterations) {
 			return 0;
 		} else {
-			return i*255/iterations;
+			return (int) (i*255.0/iterations);
 		}
 	}
